Crate mpatch

mpatch is a tool for patching files based on a matching between the source and target of the patch. Its usage is similar to Unix patch as it requires a diff file as input that specifies the changes which have been determined between two versions of the source variant. Currently, it is assumed that the diff has been calculated using Unix diff with the recommended list of arguments -Naur.

The library can be used to calculate matchings between two source files, or to apply patches read from a file or provided as text.

Examples

Patching target variants

You can patch a target variant using the apply_all function. See its documentation for more information on the paramters.

use std::path::PathBuf;
use std::str::FromStr;
use mpatch::KeepAllFilter;
use mpatch::PatchPaths;

let strip = 1;
let dryrun = true;
let matcher = mpatch::LCSMatcher;
let patch_paths = PatchPaths::new(
    PathBuf::from("tests/samples/source_variant/version-0"),
    PathBuf::from("tests/samples/target_variant/version-0"),
    PathBuf::from("tests/samples/patch.diff"),
    None,
);

if let Err(error) = mpatch::apply_all(
    patch_paths,
    strip,
    dryrun,
    matcher,
    KeepAllFilter,
) {
    eprintln!("{}", error);
}

Modules

alignment

Module for aligning patches

application

Module for applying patches

diffs

Module for types that implement reading and parsing diff files.

error

Module for error types.

filtering

Module for filtering patches

matching

Module for matching two files.

patch

Module for types and functions that represent patches and patch application.

Structs

AlignedPatch

An aligned patch contains a vector of changes that were aligned for a specific target file. The patch holds ownership of the target FileArtifact and changes it during patch application. Applying the patch consumes it to prohibit mutliple applications of the same patch to the same file. An aligned patch also has a change type that describes whether the file is created, removed, or modified.

DistanceFilter

Error

Error is the main error type of this crate and used in all high-level instances of Result<…> return values. Each error contains a message and an ErrorKind instance.

FileArtifact

Represents a file that can be patched. Each file artifact tracks the path to the file on disk and the content of the file in lines.

FileDiff

A FileDiff represents a diff between two versions of a file. Each FileDiff contains a DiffCommand (i.e., its header line), a source and a target file, and one or more hunks. Hunks contain grouped changes to lines.

FilePatch

A file patch contains a vector of changes for a specific file from a FileDiff. A file patch also has a change type that describes whether the file is created, removed, or modified.

KeepAllFilter

LCSMatcher

A simple matcher using the similar crate which offers implementations of the LCS algorithm.

Matching

A matching holds the information about lines that have been matched between a source and a target file. To this end, the matching controls two vectors of match ids: one with matchings for the lines in the source file, and one with matchings for lines in the target file. This allows for quick access to the matches by line number.

PatchOutcome

A patch outcome contains all information about a performed patch application. It contains the patched file artifact, which has already been written to disk if dryrun is disabled. Furthermore, it contains a vector of all rejected changes and the change type of the applied patch.

PatchPaths

VersionDiff

A VersionDiff represents a diff between two versions of a project or parts of a projects. A VersionDiff comprises one or more FileDiffs which in turn represent diffs for individual files.

Enums

ErrorKind

An ErrorKinds classifies which type of error has occurred.

Traits

Filter

Matcher

A trait for defining a common interface for matchers that match lines between two files.

Functions

apply_all

Applies all file patches that are found in the diff file. This function also requires a path to the directories of the source and target variants for the patch application, because it tries to match the artifacts of the source variant with the artifacts of the target variant. The calculated matching is then used to determine the best locations for a specific change to a file.