Module mpatch::diffs

Module for types that implement reading and parsing diff files.

Structs

ChangedLines

Iterator over references of HunkLines constituting line changes.

DiffCommand

A DiffCommand holds the exact call to diff used to create a FileDiff (e.g., “diff -Naur …”).

FileDiff

A FileDiff represents a diff between two versions of a file. Each FileDiff contains a DiffCommand (i.e., its header line), a source and a target file, and one or more hunks. Hunks contain grouped changes to lines.

Hunk

A Hunk consists of a source location, a target location, and one or more HunkLines. The locations describe the start and length of the changed text by line number. The source location specifies the location before the changes (i.e., the state in the source file). The target location specifies the location after the changes (i.e., the state in the target file).

HunkLine

A HunkLine contains the information about a single line in a Hunk. A HunkLine stores the text of the line, its location in the source file, its location in the target file, and its LineType.

HunkLocation

A HunkLocation defines the location of a Hunk by its line number and length in lines.

IntoChangedLines

Iterator over owned instances of HunkLines constituting line changes.

SourceFileHeader

A source file header holds the path to the source file and the timestamp of when it was read for diffing.

TargetFileHeader

A target file header holds the path to the target file and the timestamp of when it was read for diffing.

VersionDiff

A VersionDiff represents a diff between two versions of a project or parts of a projects. A VersionDiff comprises one or more FileDiffs which in turn represent diffs for individual files.

Enums

LineLocation

A LineLocation can be a RealLocation (i.e., the line actually exists in the file at that location), a ChangeLocation (i.e., the line will be added or has been removed from this location), or None (i.e., the line is not actually part of source or target). The latter is the case for EOF markings in a hunk.

LineType

Defines the type of a HunkLine (i.e., Context, Add, Remove, EOF).