Function mpatch::apply_all

pub fn apply_all(
    patch_paths: PatchPaths,
    strip: usize,
    dryrun: bool,
    matcher: impl Matcher,
    filter: impl Filter,
) -> Result<(), Error>

Applies all file patches that are found in the diff file. This function also requires a path to the directories of the source and target variants for the patch application, because it tries to match the artifacts of the source variant with the artifacts of the target variant. The calculated matching is then used to determine the best locations for a specific change to a file.

This function creates new files for files that are created by a patch, and it deletes files for file deletions in a patch, regardless of whether the lines in the patch match the lines in the file completelty.

Parameters

source_dir_path

Specifies the path to the root directory of the source variant of the diff. The source variant is the version of the program that was modified by the changes that are now to be applied.

target_dir_path

Specifies the path to the root directory of the target variant that is to be patched.

patch_file_path

Specifies the path to the diff that describes all changes that are to be applied. This diff should usually have been determined between two versions of the source variant.

rejects_file_path

You may optionally provide a path to a rejects file to which all rejected changes are written. If no path is provided, the rejects are printed to stdout.

strip

You can also define a path strip s. The strip is used to remove the leading s elements from a path (e.g., the path hello/world/directory becomes world/directory for a strip of s=1). Providing a strip is usually necessary because diffs are created based on two version of a source variant that are located in different directories (e.g., source-A/... and source-B/...). This top-level directory should be cut from the path for the patch being applied successfully.

dryrun

You should also specify whether the patch application should be made persistant (i.e., patched files are saved), or if this is only a dryrun. In case of a dryrun, the patch application is only simulated, printing all rejects to stdout without file changes.

matcher

Lastly, this function requires a matcher that is used to calculate the matching between source and target variant. See mpatch::matching for more information.