Struct mpatch::matching::Matching

pub struct Matching { /* private fields */ }

A matching holds the information about lines that have been matched between a source and a target file. To this end, the matching controls two vectors of match ids: one with matchings for the lines in the source file, and one with matchings for lines in the target file. This allows for quick access to the matches by line number.

Furthermore, a matching owns the instances of the FileArtifacts that have been matched. This ensures that the matched FileArtifacts are not altered. Note that this does not prevent the actual file being modified on disk.

Implementations

impl Matching

pub fn new( source: FileArtifact, target: FileArtifact, source_to_target: Vec<MatchId>, target_to_source: Vec<MatchId>, ) -> Matching

Creates a new Matching from the given source and target files and match id vectors.
Each line in the source and target must have an entry in the corresponding d vector at position line_number-1. The match for a line is stored as line number of its counterpart in the other file without -1 offset. This means that if the first line of both files matches, the entries of the vectors look as follows: source_to_target[0] == Some(1) target_to_source[0] == Some(1)

Note that the getter methods of the Matching struct abstract this implementation detail: matching.target_index(1) == Some(1) matching.source_index(1) == Some(1)

pub fn target_index(&self, source_index: usize) -> Option<MatchId>

Returns the match in the target file for a line number of the source file.

Input

source_index: specifies the line number of a line in the source file for which the match should be retrieved.

Output

Returns None if the source line has not been processed by the matcher. Returns Some(MatchId) if the source line has been processed. The returned MatchId is Some if there is a match in the target file; otherwise, it is None.

pub fn source_index(&self, target_index: usize) -> Option<MatchId>

Returns the match in the source file for a line number of the target file.

Input

target_index: specifies the line number of a line in the target file for which the match should be retrieved.

Output

Returns None if the target line has not been processed by the matcher. Returns Some(MatchId) if the target line has been processed. The returned MatchId is Some if there is a match in the source file; otherwise, it is None.

pub fn source(&self) -> &FileArtifact

Returns a reference to the source file instance.

pub fn target(&self) -> &FileArtifact

Returns a reference to the target file instance.

pub fn into_source(self) -> FileArtifact

Consumes this matching and returns ownership of the source file.

pub fn into_target(self) -> FileArtifact

Consumes this matching and returns ownership of the target file.