Trait mpatch::matching::Matcher

pub trait Matcher {
    // Required method
    fn match_files(
        &mut self,
        source: FileArtifact,
        target: FileArtifact,
    ) -> Matching;
}

A trait for defining a common interface for matchers that match lines between two files.

Matchers are used by mpatch to determine the alignment for a patch. This means that mpatch decides where to apply a patch based on the matching provided by a matcher.

How to implement

Ideally, a matcher implementation uses a dedicated matching algorithm to determine the differences and commonalities between two files (e.g., LCS).

A naive implementation of a matcher could iterate over the lines in both files, matching lines if they have the same content.

use std::path::PathBuf;
use std::str::FromStr;
use mpatch::{Matcher, Matching, FileArtifact};
struct NaiveMatcher;

impl Matcher for NaiveMatcher {
    fn match_files(&mut self, source: FileArtifact, target: FileArtifact) -> Matching {
        // Initialze the vectors holding the match ids
        let mut source_to_target = Vec::with_capacity(source.len());
        let mut target_to_source = Vec::with_capacity(target.len());

        // Add an entry for each line in the source and target file. Each line must have an entry in
        // the vector at position `line_number-1`.
        // The match for a line is stored as line number of its counterpart in the other file
        // without -1 offset.
        // This means that if the first line of both files matches the entries of the vectors look
        // as follows:
        // source_to_target\[0\] == Some(1)
        // target_to_source\[0\] == Some(1)
        //
        // Note that the getter methods of the Matching struct abstract this implementation detail:
        // matching.target_index(1) == Some(1)
        // matching.source_index(1) == Some(1)
        for (line_number, (source_line, target_line)) in
            source.lines().iter().zip(target.lines()).enumerate()
        {
            if source_line == target_line {
                source_to_target.push(Some(line_number));
                target_to_source.push(Some(line_number));
            } else {
                source_to_target.push(None);
                target_to_source.push(None);
            }
        }
        Matching::new(source, target, source_to_target, target_to_source)
    }
}

// Now we can use the matcher!

// Initialze some simple FileArtifacts
let file_a = FileArtifact::from_lines(
    PathBuf::from_str("file_a").unwrap(),
    vec!["SAME LINE".to_string(), "DIFFERENT LINE".to_string()],
);
let file_b = FileArtifact::from_lines(
    PathBuf::from_str("file_b").unwrap(),
    vec!["SAME LINE".to_string(), "DIFFERENT    LINE".to_string()],
);

// Call the matcher
let mut matcher = NaiveMatcher;
let matching = matcher.match_files(file_a, file_b);

// The first line matches
assert_eq!(matching.target_index(1).unwrap(), Some(1));
assert_eq!(matching.source_index(1).unwrap(), Some(1));

// The second line does not match; there is no matching for source and target
assert_eq!(matching.target_index(2).unwrap(), None);
assert_eq!(matching.source_index(2).unwrap(), None);

// There is no matching for a third line, because there was no third line in both files
assert!(matching.target_index(3).is_none());
assert!(matching.source_index(3).is_none());

Required Methods

fn match_files( &mut self, source: FileArtifact, target: FileArtifact, ) -> Matching

Determines the matching between the two fiven files. The matching takes ownership of the files to ensure that they are not changed by some other code, which would invalidate the matching, and to allow for easy access to lines depending on a match id.

Examples

The following is an example of a naive implementation that matches lines if they have the same line number and content.

fn match_files(&mut self, source: FileArtifact, target: FileArtifact) -> Matching {
    // Initialze the vectors holding the match ids
    let mut source_to_target = Vec::with_capacity(source.len());
    let mut target_to_source = Vec::with_capacity(target.len());

    // Add an entry for each line in the source and target file. Each line must have an entry in
    // the vector at position `line_number-1`.
    // The match for a line is stored as line number of its counterpart in the other file
    // without -1 offset.
    // This means that if the first line of both files matches the entries of the vectors look
    // as follows:
    // source_to_target\[0\] == Some(1)
    // target_to_source\[0\] == Some(1)
    //
    // Note that the getter methods of the Matching struct abstract this implementation detail:
    // matching.target_index(1) == Some(1)
    // matching.source_index(1) == Some(1)
    for (line_number, (source_line, target_line)) in
        source.lines().iter().zip(target.lines()).enumerate()
    {
        if source_line == target_line {
            source_to_target.push(Some(line_number));
            target_to_source.push(Some(line_number));
        } else {
            source_to_target.push(None);
            target_to_source.push(None);
        }
    }
    Matching::new(source, target, source_to_target, target_to_source)
}