Module mpatch::patch

Module for types and functions that represent patches and patch application.

Modules

alignment

application

filtering

matching

Structs

AlignedPatch

An aligned patch contains a vector of changes that were aligned for a specific target file. The patch holds ownership of the target FileArtifact and changes it during patch application. Applying the patch consumes it to prohibit mutliple applications of the same patch to the same file. An aligned patch also has a change type that describes whether the file is created, removed, or modified.

Change

A change represent a single line change (i.e., adding or removing a line of text). Each change has a content, a change type, a line number, and a change id.

FilePatch

A file patch contains a vector of changes for a specific file from a FileDiff. A file patch also has a change type that describes whether the file is created, removed, or modified.

FilteredPatch

An aligned patch contains a vector of changes that were aligned for a specific target file. The patch holds ownership of the target FileArtifact and changes it during patch application. Applying the patch consumes it to prohibit mutliple applications of the same patch to the same file. An aligned patch also has a change type that describes whether the file is created, removed, or modified.

PatchOutcome

A patch outcome contains all information about a performed patch application. It contains the patched file artifact, which has already been written to disk if dryrun is disabled. Furthermore, it contains a vector of all rejected changes and the change type of the applied patch.

PatchPaths

Enums

FileChangeType

Enum representing the three possible change types for a file: Create, Remove, and Modify.

LineChangeType

Enum representing the two possible change types for a line: Add and Remove.

Functions

apply_all

Applies all file patches that are found in the diff file. This function also requires a path to the directories of the source and target variants for the patch application, because it tries to match the artifacts of the source variant with the artifacts of the target variant. The calculated matching is then used to determine the best locations for a specific change to a file.