Module mpatch::patch::matching

Structs

LCSMatcher

A simple matcher using the similar crate which offers implementations of the LCS algorithm.

MatchOffset

Matching

A matching holds the information about lines that have been matched between a source and a target file. To this end, the matching controls two vectors of match ids: one with matchings for the lines in the source file, and one with matchings for lines in the target file. This allows for quick access to the matches by line number.

Traits

Matcher

A trait for defining a common interface for matchers that match lines between two files.

Type Aliases

MatchId

A MatchId is simply an Option<usize> where the usize is a line number in the interval [1,n].